# Use PHF for generated matches