--- a/intl_pluralrules/src/lib.rs
+++ b/intl_pluralrules/src/lib.rs
@@ -26,6 +26,7 @@
 //! ```
 #[macro_use]
 extern crate matches;
+extern crate phf;
 
 /// A public AST module for plural rule representations.
 pub mod operands;