This improves performance of language rule fetching significantly. Not as much as I'd hoped, but 5x isn't too bad.

Instead of a giant `match` (which involves a lot of string comparisons), we use phf-map, which at compile time generates an efficient hash function that we can use with the statically generated map for fast lookup.


```
# before changes
test bench_create ... bench:         604 ns/iter (+/- 73)
test bench_total  ... bench:         943 ns/iter (+/- 159)


# after changes
test bench_create ... bench:         169 ns/iter (+/- 73)
test bench_total  ... bench:         819 ns/iter (+/- 235)
```