@zbraniecki 2018-08-14
This will require `cargo regenerate-fixtures` and `cargo regenerate-fixtures-within` - https://github.com/unclenachoduh/pluralrules/blob/master/make_pluralrules/.cargo/config