@zbraniecki 2018-08-14
Contributes to #14