@unclenachoduh 2018-08-15
Thanks for the PR @Manishearth. I updated the fixtures in https://github.com/unclenachoduh/pluralrules/pull/22 so this PR can be closed.